reorganize and cleanup some imports
