+++ services/ui-src/src/utils/index.ts
@@ -13,6 +13,7 @@ export * from "./forms/forms";
 export * from "./forms/ilosFields";
 export * from "./forms/naaarPlanCompliance";
 export * from "./forms/priorAuthorization";
+export * from "./forms/standards";
 // reports
 export * from "./reports/entities";
 export * from "./reports/reports";
